Cleanup 1.3.1 merge
