+++ settings.gradle
@@ -9,6 +9,8 @@ include 'v1_12_R1'
 include 'v1_13_R2'
 include 'v1_14_R1'
 include 'v1_15_R1'
+include 'v1_16_R1'
+include 'v1_16_R2'
 include 'v1_16_R3'
 include 'v1_17_R1'
 include 'v1_18_R2'
