+++ v1_16_R2/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id 'java'
+}
+
+group 'me.codedred'
+version "${gradle.rootProject.version}"
+
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
+compileJava.options.encoding = 'UTF-8'
+
+repositories {
+    mavenLocal()
+}
+
+dependencies {
+    compileOnly project(':')
+    compileOnly 'org.spigotmc:spigot:1.16.3-R0.1-SNAPSHOT'
+}
